Separate code for dbt adaptor
